visualization_app/requirements.txt
absl-py==1.3.0
aiohttp==3.8.3
aiosignal==1.3.1
altair==4.2.0
anytree==2.8.0
asttokens==2.1.0
async-timeout==4.0.2
attrs==22.1.0
backcall==0.2.0
beartype==0.11.0
beautifulsoup4==4.11.1
blinker==1.5
bpemb==0.3.4
cachetools==5.2.0
certifi==2022.9.24
charset-normalizer==2.1.1
click==8.0.4
cloudpickle==2.2.0
colorama==0.4.5
commonmark==0.9.1
conllu==4.5.2
contourpy==1.0.6
cycler==0.11.0
Cython==0.29.28
debugpy==1.6.3
decorator==5.1.1
Deprecated==1.2.13
dill==0.3.5.1
docopt==0.6.2
entrypoints==0.4
executing==1.2.0
filelock==3.8.0
fire==0.4.0
flet==0.1.62
fonttools==4.38.0
frozenlist==1.3.3
fsspec==2022.11.0
ftfy==6.1.1
future==0.18.2
gdown==4.4.0
gensim==4.2.0
gitdb==4.0.9
GitPython==3.1.29
google-auth==2.14.1
google-auth-oauthlib==0.4.6
grpcio==1.50.0
huggingface-hub==0.10.1
hyperopt==0.2.7
idna==3.4
importlib-metadata==5.0.0
ipykernel==6.17.1
ipython==8.6.0
Janome==0.4.2
jedi==0.18.1
Jinja2==3.1.2
joblib==1.2.0
jsonschema==4.17.0
jupyter_client==7.4.6
jupyter_core==5.0.0
kiwisolver==1.4.4
konoha==4.6.5
langdetect==1.0.9
lightning-utilities==0.3.0
lxml==4.9.1
Markdown==3.4.1
MarkupSafe==2.1.1
matplotlib==3.6.2
matplotlib-inline==0.1.6
more-itertools==9.0.0
mpld3==0.3
multidict==6.0.2
multiprocess==0.70.13
nest-asyncio==1.5.6
networkx==2.8.8
numpy==1.23.4
oauthlib==3.2.1
overrides==3.1.0
packaging==21.3
pandas==1.5.0
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
platformdirs==2.5.4
plotly==5.11.0
pptree==3.1
prompt-toolkit==3.0.32
protobuf==3.20.3
psutil==5.9.4
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==10.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pydeck==0.8.0
Pygments==2.13.0
Pympler==1.0.1
pyparsing==3.0.9
pyrsistent==0.19.2
PySocks==1.7.1
python-dateutil==2.8.2
pytorch-lightning==1.8.1
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
pywin32==305
PyYAML==6.0
pyzmq==24.0.1
regex==2022.9.13
repath==0.9.0
requests==2.28.1
requests-oauthlib==1.3.1
responses==0.18.0
rich==12.6.0
rsa==4.9
scikit-learn==1.1.2
scipy==1.9.2
seaborn==0.12.1
segtok==1.5.11
semver==2.13.0
sentencepiece==0.1.95
six==1.16.0
sklearn==0.0
smart-open==6.2.0
smmap==5.0.0
soupsieve==2.3.2.post1
sqlitedict==2.0.0
stack-data==0.6.1
streamlit==1.14.0
streamlit-option-menu==0.3.2
tabulate==0.9.0
tenacity==8.1.0
tensorboard==2.11.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
termcolor==2.1.0
terminaltables==3.1.10
threadpoolctl==3.1.0
tokenizers==0.13.1
toml==0.10.2
toolz==0.12.0
torch==1.12.1
torchaudio==0.12.1
torchmetrics==0.10.2
torchvision==0.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.5.0
transformers==4.23.1
typing_extensions==4.4.0
tzdata==2022.6
tzlocal==4.2
urllib3==1.26.12
validators==0.20.0
watchdog==2.1.9
wcwidth==0.2.5
websocket-client==1.4.1
Werkzeug==2.2.2
Wikipedia-API==0.5.4
word2number==1.1
wrapt==1.14.1
xxhash==3.1.0
yarl==1.8.1
zipp==3.10.0
